add temporary routing for disjoint apps
Currently, the Initiatives editor is under heavy development and neither
the shape of its state, nor its interface with the sourcecred core, is
clearly defined just yet.

In order to get the Initiatives Editor merged in its current form, it
will be hidden behind the `/editor` URI while the new Explorer shall be
visible at the root of the dev server path.

test plan: run `yarn start --instance <instance>` and then browse to
both `localhost:8080` and localhost:8080/editor and ensure the Explorer
and Initiatives Editor load, in that respective order